prepared local/prod configs

+++ src/main/java/com/example/FlowFireHub/Utilities/DummyDataLoader.java
@@ -34,52 +34,60 @@ public class DummyDataLoader implements CommandLineRunner {
 
     @Override
     public void run(String... args) throws Exception {
-        userRepository.deleteAll();
-        flowFireRepository.deleteAll();
-        roleRepository.deleteAll();
-        chatRoomRepository.deleteAll();
+        if(dbDriver.equals("org.h2.Driver")) {
 
-        Role role = new Role();
-        role.setName("Administrator");
-        roleRepository.save(role);
+            System.out.println("==== H2 DRIVER ENV ====");
 
-        {
-            User user = new User();
-            user.setUsername("Rasmus");
-            user.setRole(role);
+            userRepository.deleteAll();
+            flowFireRepository.deleteAll();
+            roleRepository.deleteAll();
+            chatRoomRepository.deleteAll();
 
-            FlowFire fireFlow = new FlowFire();
-            fireFlow.setPassword("$2a$10$Wpn3AIuaRU4/975BYY8pLepwwV9phOah65k6uh3Kk2rAJn3ghq4Li");
-            fireFlow.setUsername("Rasmus");
-            fireFlow.setUser(user);
-            flowFireRepository.save(fireFlow);
+            Role role = new Role();
+            role.setName("Administrator");
+            roleRepository.save(role);
+
+            {
+                User user = new User();
+                user.setUsername("Rasmus");
+                user.setRole(role);
+
+                FlowFire fireFlow = new FlowFire();
+                fireFlow.setPassword("$2a$10$Wpn3AIuaRU4/975BYY8pLepwwV9phOah65k6uh3Kk2rAJn3ghq4Li");
+                fireFlow.setUsername("Rasmus");
+                fireFlow.setUser(user);
+                flowFireRepository.save(fireFlow);
+            }
+
+            {
+                User user = new User();
+                user.setUsername("Peter");
+                user.setRole(role);
+
+                FlowFire fireFlow = new FlowFire();
+                fireFlow.setPassword("$2a$10$Wpn3AIuaRU4/975BYY8pLepwwV9phOah65k6uh3Kk2rAJn3ghq4Li");
+                fireFlow.setUsername("Peter");
+                fireFlow.setUser(user);
+                flowFireRepository.save(fireFlow);
+            }
+
+            {
+                ChatRoom room = new ChatRoom("general", false);
+                chatRoomRepository.save(room);
+            }
+
+            {
+                ChatRoom room = new ChatRoom("synthetik", false);
+                chatRoomRepository.save(room);
+            }
+
+            {
+                ChatRoom room = new ChatRoom("offtopic", false);
+                chatRoomRepository.save(room);
+            }
         }
-
-        {
-            User user = new User();
-            user.setUsername("Peter");
-            user.setRole(role);
-
-            FlowFire fireFlow = new FlowFire();
-            fireFlow.setPassword("$2a$10$Wpn3AIuaRU4/975BYY8pLepwwV9phOah65k6uh3Kk2rAJn3ghq4Li");
-            fireFlow.setUsername("Peter");
-            fireFlow.setUser(user);
-            flowFireRepository.save(fireFlow);
-        }
-
-        {
-            ChatRoom room = new ChatRoom("general", false);
-            chatRoomRepository.save(room);
-        }
-
-        {
-            ChatRoom room = new ChatRoom("synthetik", false);
-            chatRoomRepository.save(room);
-        }
-
-        {
-            ChatRoom room = new ChatRoom("offtopic", false);
-            chatRoomRepository.save(room);
+        else {
+            System.out.println("==== NO H2 ====");
         }
 
     }
